move HeadVec to wasmi_collections
